Completed logging subsystem, started docs
Completed a majority of the Logging subsystem to create Loggers. Created
LogSinks for printing Logs to stdout on POSIX platforms, and created a
LogService component as part of the standard collection to implement a
reference Logger. Macros will interact with Loggers such as LogService.
Reference FSW illustrates usage of the logger. In addition wrote very
simple docs describing the work so far.